# frontend/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Annotate Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/work-items/{item_id}/context": {
      "get": {
        "summary": "Work Context",
        "operationId": "work_context_api_v1_work_items__item_id__context_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/draft": {
      "put": {
        "summary": "Save Draft",
        "operationId": "save_draft_api_v1_work_items__item_id__draft_put",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/submit": {
      "post": {
        "summary": "Submit Annotation",
        "operationId": "submit_annotation_api_v1_work_items__item_id__submit_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/review": {
      "post": {
        "summary": "Review Item",
        "operationId": "review_item_api_v1_work_items__item_id__review_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/clear": {
      "post": {
        "summary": "Clear Annotations",
        "operationId": "clear_annotations_api_v1_work_items__item_id__clear_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/data": {
      "get": {
        "summary": "Item Data",
        "operationId": "item_data_api_v1_work_items__item_id__data_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/media/{media_key}": {
      "get": {
        "summary": "Item Media",
        "operationId": "item_media_api_v1_work_items__item_id__media__media_key__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "media_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Key"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "available",
          "annotation_assigned",
          "annotating",
          "review_pending",
          "review_assigned",
          "reviewing",
          "changes_requested",
          "completed",
          "cancelled"
        ],
        "title": "ItemStatus"
      },
...
      "QaStatus": {
        "type": "string",
        "enum": [
          "unchecked",
          "passed",
          "rejected"
        ],
        "title": "QaStatus"
      },
...
      "ReviewInput": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(approve|request_changes)$",
            "title": "Decision"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewInput"
      },
      "RevisionInput": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "segments.v1"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "base_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision Id"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "RevisionInput"
      },
...
      "TaskItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "claim_order": {
            "type": "integer",
            "title": "Claim Order"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "annotator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotator Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "qa_status": {
            "$ref": "#/components/schemas/QaStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "package_id",
          "episode_id",
          "claim_order",
          "status",
          "annotator_id",
          "reviewer_id",
          "qa_status",
          "updated_at"
        ],
        "title": "TaskItemOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkContext": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/TaskItemOut"
          },
          "episode_index": {
            "type": "integer",
            "title": "Episode Index"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "fps": {
            "type": "number",
            "title": "Fps"
          },
          "tasks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tasks"
          },
          "data_url": {
            "type": "string",
            "title": "Data Url"
          },
          "video_urls": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Video Urls"
          },
          "latest_revision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Revision"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "quality_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Comment"
          }
        },
        "type": "object",
        "required": [
          "item",
          "episode_index",
          "length",
          "fps",
          "tasks",
          "data_url",
          "video_urls",
          "latest_revision"
        ],
        "title": "WorkContext"
      },
...
    }
  }
}